Test: split out foreign utils
